requirements.txt
awscli==1.27.13
boto3==1.26.17
botocore==1.29.17
colorama==0.4.4
docutils==0.16
Faker==15.3.3
greenlet==2.0.1
jmespath==1.0.1
numpy==1.23.5
pandas==1.5.2
psycopg2-binary==2.9.5
pyasn1==0.4.8
python-dateutil==2.8.2
pytz==2022.6
PyYAML==5.4.1
rsa==4.7.2
s3transfer==0.6.0
six==1.16.0
SQLAlchemy==1.4.44
urllib3==1.26.12
